requirements.txt
flask
seleniumbase
Flask-JWT-Extended
pycryptodome